requirements.txt
beautifulsoup4
lxml
requests
schedule
python-dotenv